+++ src/main/kotlin/graphql/dataTransferObjects/CannotDeleteAccount.kt
@@ -2,6 +2,6 @@ package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
 import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 
-class CannotDeleteAccount {
+object CannotDeleteAccount {
     fun getPlaceholder() = Placeholder
 }