+++ src/main/kotlin/graphql/dataTransferObjects/DeletedPrivateChat.kt
@@ -1,5 +1,5 @@
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
-class UpdatedGroupChatPic(private val chatId: Int) : ChatsSubscription, GroupChatMetadataSubscription {
+class DeletedPrivateChat(private val chatId: Int) : ChatsSubscription {
     fun getChatId(): Int = chatId
 }