+++ src/main/kotlin/graphql/routing/Subscriptions.kt
@@ -48,7 +48,7 @@ fun routeGraphQlSubscriptions(context: Routing) {
         ),
         GraphQlSubscription(path = "accounts-subscription", "subscribeToAccounts", isAuthenticated = true),
         GraphQlSubscription(path = "chat-accounts-subscription", "subscribeToChatAccounts", isAuthenticated = false),
-        GraphQlSubscription(path = "group-chats-subscription", "subscribeToGroupChats", isAuthenticated = true),
+        GraphQlSubscription(path = "chats-subscription", "subscribeToChats", isAuthenticated = true),
         GraphQlSubscription(
             path = "group-chat-metadata-subscription",
             "subscribeToGroupChatMetadata",